+++ src/v2/view-builder/views/ov/ChallengeOktaVerifyFastPassView.js
@@ -1,11 +1,9 @@
-import { $, loc, createCallout } from 'okta';
+import { $, createCallout, _ } from 'okta';
 import { BaseFormWithPolling } from '../../internals';
 import Logger from '../../../../util/Logger';
-import { CHALLENGE_TIMEOUT, 
-  OV_UV_ENABLE_BIOMETRICS_FASTPASS_DESKTOP, 
-  OV_UV_ENABLE_BIOMETRICS_FASTPASS_MOBILE } from '../../utils/Constants';
+import { CHALLENGE_TIMEOUT } from '../../utils/Constants';
 import BrowserFeatures from '../../../../util/BrowserFeatures';
-import { doChallenge } from '../../utils/ChallengeViewUtil';
+import { doChallenge, getBiometricsErrorOptions } from '../../utils/ChallengeViewUtil';
 
 const request = (opts) => {
   const ajaxOptions = Object.assign({
@@ -127,36 +125,14 @@ const Body = BaseFormWithPolling.extend(Object.assign(
     },
 
     showCustomFormErrorCallout(error) {
-      const errorSummaryKeys = error?.responseJSON?.errorSummaryKeys;
-      const isBiometricsRequiredMobile = errorSummaryKeys 
-          && errorSummaryKeys.includes(OV_UV_ENABLE_BIOMETRICS_FASTPASS_MOBILE);
-      const isBiometricsRequiredDesktop = errorSummaryKeys 
-          && errorSummaryKeys.includes(OV_UV_ENABLE_BIOMETRICS_FASTPASS_DESKTOP);
-
-      if (isBiometricsRequiredMobile || isBiometricsRequiredDesktop) {
-        const bulletPoints = [
-          loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point1', 'login'),
-          loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point2', 'login'),
-          loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point3', 'login')
-        ];
-
-        // Add an additional bullet point for desktop devices
-        if (isBiometricsRequiredDesktop) {
-          bulletPoints.push(
-            loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point4', 'login')
-          );
-        }
-
-        const options = {
-          type: 'error',
-          className: 'okta-verify-uv-callout-content',
-          title: loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.title', 'login'),
-          subtitle: loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.description', 'login'),
-          bullets: bulletPoints,
-        };
-        this.showMessages(createCallout(options));
-        return true;
+      const options = getBiometricsErrorOptions(error);
+      
+      if (_.isEmpty(options)) {
+        return false;
       }
+
+      this.showMessages(createCallout(options));
+      return true;
     },
   },
 ));
