+++ src/v2/view-builder/views/ov/ChallengeOktaVerifySSOExtensionView.js
@@ -1,8 +1,7 @@
 import BaseAuthenticatorView from '../../components/BaseAuthenticatorView';
 import { BaseForm } from '../../internals';
-import { loc, createCallout } from 'okta';
-import { OV_UV_ENABLE_BIOMETRICS_FASTPASS_DESKTOP, 
-  OV_UV_ENABLE_BIOMETRICS_FASTPASS_MOBILE } from '../../utils/Constants';
+import { loc, createCallout, _ } from 'okta';
+import { getBiometricsErrorOptions } from '../../utils/ChallengeViewUtil';
 
 // for EA,
 // redirect is needed for Apple SSO Extension to intercept the request, because
@@ -32,36 +31,14 @@ const Body = BaseForm.extend({
   },
 
   showCustomFormErrorCallout(error) {
-    const errorSummaryKeys = error?.responseJSON?.errorSummaryKeys;
-    const isBiometricsRequiredMobile = errorSummaryKeys 
-        && errorSummaryKeys.includes(OV_UV_ENABLE_BIOMETRICS_FASTPASS_MOBILE);
-    const isBiometricsRequiredDesktop = errorSummaryKeys 
-        && errorSummaryKeys.includes(OV_UV_ENABLE_BIOMETRICS_FASTPASS_DESKTOP);
-
-    if (isBiometricsRequiredMobile || isBiometricsRequiredDesktop) {
-      const bulletPoints = [
-        loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point1', 'login'),
-        loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point2', 'login'),
-        loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point3', 'login')
-      ];
-
-      // Add an additional bullet point for desktop devices
-      if (isBiometricsRequiredDesktop) {
-        bulletPoints.push(
-          loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.point4', 'login')
-        );
-      }
-
-      const options = {
-        type: 'error',
-        className: 'okta-verify-uv-callout-content',
-        title: loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.title', 'login'),
-        subtitle: loc('oie.authenticator.oktaverify.method.fastpass.verify.enable.biometrics.description', 'login'),
-        bullets: bulletPoints,
-      };
-      this.showMessages(createCallout(options));
-      return true;
+    const options = getBiometricsErrorOptions(error);
+    
+    if (_.isEmpty(options)) {
+      return false;
     }
+
+    this.showMessages(createCallout(options));
+    return true;
   },
 });
